backend/requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.15
django-allauth==0.45.0
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-timezone-field==4.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
kombu==5.2.4
oauthlib==3.1.1
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.1
pycparser==2.20
pytest-django
PyJWT==2.4.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.1
redis==4.2.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
vine==5.0.0
wrapt==1.14.0
factory_boy==3.2.1
